implied: fix a bug handling partial implications
The problem with partial implications is that for the gate_sm, then
smatch_extra.c can give us the exact states.  When we do this, it ends
up throwing away some implications, but we were already in trouble and
taking too long to process things so there was no way to *use* the
implications so no harm done.

I decided that instead of returning partial implications it was probably
better to just return NULL.

Also the recurse limit was not being used consistently.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>