math: fix RL_ABSOLUTE for unknowns
In the get_rl_sval(), I accidentally changed it to return false for
RL_ABSOLUTE.  It's the code that does:

	if (!ret)
		return false;

When I remove that check, that means "ret" isn't used any more and so I
removed it completely.  I also added a check to never set *res when
RL_EXACT is set.  I don't think that check will ever trigger, but it
prevents a potential crash (and replaces it with a more subtle math
error).

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>