mtag: update the mtag when it's set using a pointer
Say we have:

	int *p = &global;

	*p = 42;

Then the original code wasn't updating global to say that it could be 42.
We need to look up the implied value of *p and find that it's an mtag and
update the mtag.

I had to create the get_implied_value_low_overhead() function because
creating fake cur_stree() stuff, generates more states and causes us to do
more mtag updates.  So we end up entering a forever recursive loop.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>