param_used: don't save bogus data
I've added __in_fake_parameter_assign to the list of ignored situations.
The downside of this is that it's possible we miss certain use cases, but
that's probably fine because we really filter unused data very sparingly.

The more important change is that I added __in_function_def to the list of
ignored situations.  Recently I changed smatch_extra.c to look up unmatched
states so that if we change/limit a global variable on one path, we look
up the mtag for the unmatched path.  That also changes how locals and
function arguments are handled.  So now if one caller passes in data and
the other callers don't we look it up and it gets recorded as used.

It's not possible to use and data while we're still processing the function
definition so this change should be totally harmless.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>