math: handle __builtin_offsetof() better
Most of the time __builtin_offset() is a known value and that's simple
enough because Sparse evaluates it and turns into a EXPR_VALUE.  But
if it's an offset into an array, and the array index isn't known then
that's tricker.  There is some bpf code which uses this in the kernel so
this patch silences a false positive.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>