integer_overflow: split some integer overflow handling into its own file
This isn't really code which is ready for use...  It's just that I wrote
it three years ago and I move the can_integer_overflow() function to this
file from smatch_math.c.  And it's really annoying when I look at the
difference between the published code and my code to see that diff all the
time...

I've tried a lot of different experiments to find integer overflow bugs but
it's a very hard nut to crack.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>