mtag: push some mtag seed code (but don't enable it)
The memory tag code is my new thing I'm working on.  I have an idea vaguely
how it works, but I haven't totally figured out the details.

The idea with memory tags is that normally people allocate a junk of memory
and they have a pointer to the start of it.  Like if you have a file scope
struct:

struct my_type whatever {
	int type, max, a, b, c;
	int (*frob)(struct my_type *p);
	...
};

That's very simple and the pointer to that is &whatever.  Anyway, the start
of a memory allocation is called an mtag in Smatch...

Then in the kernel you have a bunch of different ways to allocate memory
dynamically with kmalloc() or kzalloc() or a bunch of other functions.  The
dynamically allocated memory is a lot hard to track.

The reason that we're tracking these is so that we can tie function
pointers to the data that gets passed to them.  So we're really only
interested in mtags that have a function pointer.  The others are
already tracked in other ways.

In the kernel, there are a few places which generate interesting mtags.
The main one is probe() functions.  So we want a driver to know when it's
talking about it's own driver specific data.  So we pass an MTAG_SEED to
the probe function.  This is done with manual annotation.  The probe
function immediately generates a function specific mtag and replaces the
seed with it's own mtag.

Then the other situation is that you have a subsystem allocation function
like:  "driver_ptr = alloc_subsystem_stuff();".  The same kmalloc() call
is used to allocate space for every driver.  In that situation, we end up
generating an alias in the caller.  The code to do this correctly is not
written yet.  I think we need to pass a separate state which can only be
passed as a return_states to mark an allocation as new.  You can't have an
alias of an alias.  The state of the returned mtag is "fully known" when we
return it, and then all the callers generate their own mtags.

Anyway, these are the basic rules in my head for how mtags get created for
caller_info (mtag seeds) and return_states (aliases).

This code is part of that work, but I haven't put any of the hooks in place
to use it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>