mtag: tweak how get_array_mtag_offset() works
The main thing here is that I'm looking at how get_mtag() works and trying
to get rid of it, or rename it or something.

But say you have "foo->bar[4];" then "foo" is the mtag, and we have to
calculate the number of bytes from foo to ->bar[4].  The number of bytes
from foo to ->bar[0] is start_offset.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>