conditions: set parent expression or bitwise ORs in conditions
Sometimes people use bitwise ORs in condition expressions.  I guess maybe
it is a speedup??  Anyway, I wasn't setting the parent expression in that
case.

I was hoping this would fix the validation test for precedence, but it
doesn't.  Never the less, it seems like the right thing to do.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>