address: fix handling of unknown &foo->bar expressions
Sometimes you have an expression "&foo->bar" where "foo" is unknown, and
"bar" is not the first member of the struct.  We want to be to marked as
a valid pointer.  But sometimes get_implied_rl() return the whole range
list and so there really isn't anything implied about it...  So we should
fall back to the "valid pointer range list" code.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>