uninitialized: treat "if (ret < 0) {" like "if (ret) {"
There are quite a few times where Smatch doesn't know what the exact
return values are for a function.  And then you have places where the
mixes error checks like "if (ret) {" and "if (ret < 0) {".  When
Smatch sees the "if (ret < 0) " it generates an error that the parameters
could be uninitialized when the return is a non-zero positive value.
But in the kernel the two error conditions are almost always
interchangeable so now Smatch just treats them that way for this
uninitialized variable check.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>