bits: track which bits are set
Say you have a condition:

	if (x & (1 << 7)) {

Then you know that BIT(7) is set.  That's often useful information to know.
Also it's sometimes nice to know if the bottom bits are clear.  Anyway,
smatch_bits.c is there to track which bits might be set and which
definitely are.

In the future it might be nice to know that only one bit is set and other
information, but this code is already very bad at its less ambitious
goals...

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>