ranges: improve handling of param comparisons
There are a couple ways that you can express that a return is equal to a
parameter "s32min-s32max[==$1]" and "s32min-s32max[$1]".  The first way
would be if we returned a variable which was equal to $1 and the second way
is if we return the actual $1 variable itself.  There is effectively no
difference, but it's handled by different subsystems.  I didn't plan
it that way, but I sort of find the format nicely expressive.

Anyway, the problem here was that user_data uses the ==$1 information.  Say
a function returns USER_DATA s32min-s32max[$1], and we know what $1 is in
the caller, then we can limit the USER_DATA in the caller.  But because of
this bug it wasn't handling s32min-s32max[$1], it was only handling [==$1].

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>