kernel_user_data: fix handling of "*$" parameter types
The idea here was that, if we pass a user controlled buffer "p" where p is
a void pointer, then it would make *p as s64min-s64max user controlled.
The problem is that the "type == void_ctype" check was reversed so it
just messed up known pointer types and left void pointer alone.

This code isn't required.  Void pointers are handled as expected if we just
delete it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>