db: fix a struct member test in caller info
What the code is doing here is that it's looking at "frob(p->foo);" and
if we have a state "p->foo->one" then it changes it to "$->one".  But if
we have a state "p->foobar" then we don't want to change it to "$bar".
The bug is how underscores are handled.  If we had a state "p->foo_bar"
then it was changed to "$_bar".

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>