db: return absolute range list instead of whole_rl
The bug here was what if you have:

	return u16val & mask;

We know the return value has to be less than u16_max, but in the original
code we were returning s32min-s32max (the function returns an int).

Also it feels weird intuitively but on my test case get_implied_rl() was
returning s32min-s32max.  I'm tempted to make get_implied_rl() return false
if it's just going to give useless whole_rl information.  I have held off
on that because I would need to update the validation/ directory and
because I am a scared-ie-cat developer instead of a bold 10x hero coder.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>