user_data: clear states on PARAM_SET
We should clear the state when we hit a PARAM_SET.

This code was written really early on and shows...  Normally we would use
the modification_hook() to clear the user data.  I probably should do that
here.

But the old code only cleared the states for assignments and that basically
works.  I've added PARAM_SET and that pretty much covers everything except
foo++.  I don't want foo++ to clear user data.  Also using the modification
hooks means you have to get the order correct so it's slightly more
complicated to implement.

We can revisit this later if we want.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>