user_data: don't set the returned state if the param was not modified
I'm going to commit this without even testing it because it just feels so
right.  A parameter can't suddenly be user data unless it has been
modified.

Btw, we track two ways of setting user data.  One is if you assign
parameters to each other "*param1 = param2;" this is USER_DATA and the
other way is if we copy it directly from the user like:
"copy_from_user(param1, src, size)" this is USER_DATA_SET.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>