kernel_user_data: basically revert "fix handling of "*$" parameter types"
I misunderstood what the code was doing when I wrote commit b09bb5aa2ad4
("kernel_user_data: fix handling of "*$" parameter types").  Say we have
a function that passes skb->data to a function which takes a certain type
of struct pointer, then in the caller_info we have *$ is user data but in
the function itself we want to say that foo->bar, foo->baz etc are user
data.  (We could do this in the caller instead but the code would be more
complicated and it ends up being the same either way).

But one thing that we don't want is to preserve that *foo is user data and
the original code did that.  So now we only look for known struct types not
char arrays or void pointers.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>