kernel_user_data: track capped information this module
Before, there was smatch_capped.c and smatch_kernel_user_data.c and they
were independent.  What smatch_capped.c does is it says that this variable
has been capped at an *unknown* limit.  Anyway the problem is that say
we merge a capped user data with uncapped non-userdata then the result is
uncapped user data.

In olden times I could have handled this by just marking it as non-user
data but with Spectre bugs then we have to track user data.  The solution
is that we have to track the capped information along with the user estate.
This will actually be useful for detecting array underflow bugs and also
integer overflow bugs as well so it's probably the right thing generally
even though it was more complicated to write the code.

The format in the DB would look something like: "s32min-s32max[==$1,c]"
where the "c" would indicate "capped".  I think we will end up moving more
and more attributes from the estate into the square brackets like that.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>