ranges: fix AND handling
I didn't understand what this code was doing at first and I copied it
badly.  I assumes that zero is always possible, then the next lowest
possible value is when only the lowest bit is set on one of the range
lists.  So we need to adjust the minimum before we take the intersection.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>