db/kernel.return_fixes: add a bunch of functions
I moved the dma_fence_timeout stuff around.  One side note with those
functions is that they return a long and smatch often thinks that it could
be larger than INT_MAX.  A bunch of places store the return in a int.  I
guess they verified that the timeout is less than INT_MAX?  Anyway, I just
always pretend that it returns an int for simplicity.

I moved some get_user_pages from fixup_kernel.sh because it can be handled
here and then all the code is together in the same place.

The other thing is that array_index_mask_nospec() returns either zero or
a mask based on the size of the array.  But it's in assembly and Smatch
doesn't parse assembly.  This patch turns it into a no-op.  People
shouldn't rely on the nospec stuff to do their bounds checking.  Except
for, it does probably work very to prevent array underflows.  If people
start relying on it for that then I'll have to add a more complicated
hack around.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>