kpatch.sh: update to use git commit
When I first started writing patches I didn't have git installed because
my internet was too bad to download the whole kernel git archive.  I've
been using `git commit` for several months now, and it's not really ideal
for my workflow but these days it is the law of the land.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>