container_of/mtag: handle more advanced containers
Before we could store something when a function was called like
"foo->bar(foo);" but now we can go further back "foo->bar->baz(foo);".
Also if we have saved a pointer to the argument
"p = foo->bar; foo->baz(p);" then that's handled as well.

There is quite a bit more thinking that needs to be done here...  I'm not
totally sure the callers are able to use this information yet.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>