function_hooks/assigned_expr: handle fake return assignments better
We have a bunch of functions that do:

void *get_data(struct whatever *p)
{
	return p->foo.driver_data;
}

In the caller, we have:

	data = get_data(p);

But in the assigned expression we want to save that we are doing:

	data = p->foo.driver_data;

So we have to ignore the function assignment and only use the fake
assignment.  Also I improved the fake assignment to deal with temporary
variables better.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>