ranges: handled truncation better in show_rl()
In the old code, when show_rl() didn't fit into the 255 char buffer then
it just tacked a '+' character on to the end.  That wasn't handled very
well in the perl DB code which converts from a string back to numbers.

Now we just end in a "-type_max" instead.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>