ranges: collapse mtag pointer range lists when they get too large
With the mtag code, what happens is that we keep merging mtags together
until they list is too long to be stored in the DB (I don't store more
than 255 characters in the DB for range lists).  Then it becomes a kind of
nonsense incomplete range list.

It's better to just change those types of range lists into the valid
pointer range list.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>