kernel_user_data: improve how returning user data is handled
In the original code, if we found "*$" was user data then we didn't check
whether just "$" was user data...  We need to look for three things and
there is no point in short circuiting the process if we find one.

We first search the states for if "$->bar" has a state and it's user data.
Then we search for if we're have "return ntohl(foo);" which doesn't set a
state but is still user data.  Then we check for if we "return skb->data;"
which doesn't have a state but points to user data.

We obviously don't want to record the same things twice so we set a flag
if "return_found" or "pointed_at_found" but otherwise we check all three.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>