kernel: modify build via env
For certain kernel builds, additional parameter might be specified
that are not picked up by the build system of the kernel via the
environment. This change allows to modify the build target, as well
as extra build arguments for the kernel build.

The following two environment variables are used to control this
behaviour. The value of the target variable is overriden in case
the environment variable is specified.

SMATCH_ENV_TARGET target to be build
SMATCH_ENV_BUILD_PARAM other parameters to be forwarded to make

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>