math: fix and improve abs() handling
The start of this problem is that -s32min is still s32min as opposed to
s32max because of integer overflows.  So then when you do abs(unknown) then
really it should be type_min,0-type_max.  But most of the time, probably
the value isn't going to be type_min, we just don't know what it is so
most of the time we should assume that abs() works as intended.

So now what I've done is that I've said, let's assume abs() gives you
positives unless the unknown variable comes from the user and in that case
let's assume that maybe they maliciously chose type_min.

Also instead of just converting the rl_min/max let's convert each value in
the rl.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>