container_of: fix the validation test case (bit of a hack)
The sm_tag1.c test code looks like this:

	struct foo *p = &one_struct;

	ret = p->func(p);

But we're using the assigned expressions so it ends up looking like:

	p->func(&one_struct);

So the relationship is not obvious.  Really the common case is that we
don't need to look at assigned expressions.

But this still feels like a hack.  Perhaps what we need is a
map_long_to_short() function that returns expression pointers instead of
var/syms.  I'm not totally sure.  I will almost certainly revisit this code
in the near future.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>