user_data: pointer math is generally not user data
The story here is that I was trying to debug a false positive, and I
decided to solve it my making pointer subtraction into a special case.
Smatch has trouble with subtraction generally so when you're dealing with
pointers probably the kernel has already checked.

I kind of made this code as least invasive as possible because it's a bit
of a hack.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>