user_data: improve user_rl_capped()
The real bug fix here is that it wasn't calling strip_expr() so binops
with parentheses were not parsed correctly "(a + b)".

I also removed the is_capped() check because it's probably not required and
it has a bit of subtle semantics.  I added a get_value() at the start
because get_value() is very fast.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>