err_ptr, core: fix how pointers with -m32 (passing zero to PTR_ERR())
When you compiled the with -m32 then it generated tons and tons of false
positives that you were passing zero to PTR_ERR().  Unfortunately, I rarely
compiled code that way and for legacy reasons I had -m64 hard coded into
my scripts so it was hard for me to reproduce this warning so it took a
while to figure out what was going on.

The bug is that the code was using ULONG_MAX as ptr_max instead of parsing
-m32/-m64 first.  That parsing happens in sparse_initialize(argc, argv, &filelist).
I wanted it to happen before we registered the checks so I moved that to
main() in smatch.c instead of smatch() in smatch_flow.c.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>