param_used: use __get_state() instead of get_state() for internals
I wrote a check which used PARAM_USED data to print an error if we were
using invalid data.  It turned out that PARAM_USED is probably not enough
of a use to trigger a warning.  But the other thing that I found was that
we were setting PARAM_USED way too much and that results in a larger than
necessary DB.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>