extra: handle += better
There was a bug handling "foo += bar;" and then I looked at the code to
handle it and I decided to just re-write it instead of trying to understand
it.

The only tricky thing about the "+= bar" code is that if we're inside a
loop then I want to go up to s32max.  Otherwise it's handled exactly the
same as all the other "*= bar" and "/= bar" etc.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>