return_to_param: re-write some stuff
The truth is that I'm not terribly sure this patch is correct...  It did
fix a bug, but did it introduce some new stuff?  I don't know.  I didn't
notice any new warnings when I compiled the kernel.

Anyway, I've been running it for a while and also I feel like it makes the
code simpler so I'm committing it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>