function_hooks:  fix a bug calling return_implies hooks
The original code really expected that the return_implies hook would
match the DB information exactly.  But quite often that's not the case.
For example, the DB might thing failures can be s32min-(-1) or (-4094)-(-1)
or s32min-(-1),1-s32max or just (-12).  So anyway, trying to match exactly
is almost impossible and it's crazy to try because the error codes can
change in the future.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>