user_data: Fix handling of inline functions
Before we used to run a select query on the DB but the query didn't handle
inlines properly.  It's better to just store the returned state as
"return %p" like other place do in the code.  It means that multiple states
get merged together nicely.

In the original code if the returned state is an assigned state then this
was already what happened, but if it was something like
"__smatch_user_rl(returns_user_data());" or "return returns_user_data();"
and it was inline then it wasn't handled properly.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>