kernel_user_data: more capped tweaking
I'm not really sure what I'm doing with trying to handle subtraction.  :P
Smatch is not yet good enough at doing comparisons to handle subtractions
with enough accuracy.  This marks a bunch more stuff as capped.

If we mask or do a modulo then that's another way of capping the data.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>