db/fixup_kernel.sh: update how kmalloc()/vmalloc() are handled
In the old code didn't split things up in a very useful way.  It should be
split up by success/fail.  This results in that the callers also tend to
split up their returns by success/fail.

This matters more than it should I think.  I can't totally recall why now
though because I was working on this before the vacation...

Also the kmalloc_slab() and slab_alloc_node() returns are out of date
now that pointers are unsigned.  The kmalloc_large() stuff doesn't make
a ton of sense or seem to be required any more.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>