ranges: re-write rl_intersection()
The bug is here is that when are using mtags, then doing an rl_invert()
is problematic.  Say we have a bunch mtags, then we invert it and it
creates a kind of ugly looking range list.  But once a range list gets
too complicated then we start to truncate it in various ways.  So if the
mtag list is on the border case, the inverted list has two extra dranges
so maybe we truncate it and lose some information.

The end result is that rl_intersection returns a NULL (for the empty list)
which is obviously wrong.

The way that rl_intersetion() works is not intuitive at all, so I've just
re-written to work in a more intuitive way and it avoid inverting
altogether.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>