db/smdb: add info command with type filtering
Add a new command to smdb that shows caller info filtered by a
user specified type. This allows us to remove the abandoned,
broken and undocumented user_data and param_value commands.

It's also possible to remove the unamed command (i.e. calling
smdb with just a function), as this new 'info' command provides
the same functionality when a user does not specify a type. However
let's keep it for now to avoid confusing users.

Signed-off-by: Andrew Murray <andrew.murray@arm.com>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>