ranges: fix truncates_nicely()
The truncates_nicely() function theory is that if you have a range list
10000000001-10000000002 and you truncate away everything greater than 10
then the resulting range list is 1-2.  That's a nice truncate.  But if you
have 10000000001-20000000002 then the resulting range list is 0-10.
Unknowns are not nice.  So for char you would want to say if
((min & ~0xff) == (max & ~0xff)) then it truncates nicely.

The shifts were wrong because it should have been "bits" instead of
"64 - bits".  I guess I only tested this with int before.  Also the right
shift is harmless but it isn't really required (I think/hope!).

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>