extra/comparison: fix really horrible bug handling assignments
This is really grisly.  So what was happening here was that I had:

	foo->bar = unknown_char;

But then it would look up "foo->bar" and get the information from the DB
and it there was another bug which was also wrong information which was
that foo->bar was zero.  And it said that "unknown_char" must also be zero
so this was "foo->bar = 0;"...  Anyway, really weird and terrible and I
can't believe it took this long to notice the bug.  :(

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>