mtag: ignore when set assign structs (only look at struct members)
When we do:

	copy_from_user(to, src, size);

What happens is that Smatch does a "*to = *src;" assignment.  So then
we end up setting "*to = s64min-s64max;".  This is very useful for
tracking user data.

The problem is that when we're tracking the mtag then the first struct
member "to->first_member;" has the same mtag and zero offset as *to.  There
are several fixes.  First don't save the wrong mtag data, second don't
look for it, and third if you get insane data then ignore that.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>