param_to_mtag_data: don't save assignments to local variables
Now that Smatch is generating mtags for stack variables, it ends up that
we save tons and tons of these mtag assignments in the return_states table.
They're not really useful, so I'm going to filter them out.

This patch is likely not going to be the last tweak.  I don't like the
is_local_variable() function.  Probably this is the correct way to
implement it.  Also it's worth considering what gets stored in mtag data
for local variables.

But those are big complicated questions and this patch is small and sort of
obvious.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>