math: undo recent change to handle_variable()
The original patch which changed this condition wasn't based on fixing any
bug, it was just something I noticed when reviewing the code.  Since then
this change has exposed some existing bugs and caused new bugs.

One new bug is if we want to update the mtag data to say that a variable
is unknown, then in the new code it doesn't like unknowns so it looks up
the mtag data and updates back to the existing mtag data.

I could work around that, but I don't think the estate_is_whole() path is
required and I'm not sure it's the correct thing from a philosophical
point of view either.  The first time we encounter a variable that's not in
smatch extra we look up the mtag information and that is enough.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>