smdb.py: fix param tracking now that we use $ instead of p
Since d8084fee2b7c the data source of a parameter is now stored with the
notation: $0 instead of p0 (where the $ or p indicates that the data source
of the parameter is from the callers function parameters, the value next to
it is the parameter number).

Let's update the smdb script to reflect this.

Fixes: d8084fee2b7c ("data_source: change how parameter data source information is recorded")
Signed-off-by: Andrew Murray <andrew.murray@arm.com>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>