ranges: fix bitwise AND handling
The code to handle bitwise was totally wrong.  One thing it was doing was
rl_intersection() so it often said that there was no intersection and the
only possible result was zero.  0xf00f and 0x000f have no shared values but
they have shared bits so the original code doesn't make sense.

The right way to handle this is to just re-use the code in smatch_bits.c.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>