extra: save fake assignement information
The problem is say you have:

	foo = bar;

That sets "foo->baz = bar->baz;", but imagine here that we have do have
any information about either of them.  In the past, we wouldn't bother to
save anything in smatch_extra.c, but unfortunately that has become
problematic now.

The issue is that smatch_mtag.c is storing other assignments where we set
"foo->baz = 0;" so now smatch_absolute.c thinks foo->baz is zero instead of
unknown.

So now we have a couple potential solutions.  We could add an exception in
smatch_mtag_data.c.  That might mean that later we would have to look up
the data in the in_mem DB as well as the on disk DB.  It's not totally
clear what the rules would be.  The other option is to just store the
state in smatch_extra.c.  Since we're already storing stuff either way we
might as well do it the simplest, cleanest way.

One thing to note is that we're still not saving information for when
param_set is unknown.  That code was based on testing and trying to speed
things up.  I think this code was based just on noticing that we could
save some memory.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>