param_used: use __get_state() in some places instead of get_state()
The param_used stuff is supposed to make the DB smaller by not passing
data which is not used.  It looks if if get_state() is used, but there are
a bunch of places where get_state() is used automatically just based on
information in the DB and not based on what the code looks like.  So we
end up saying that if it's in the DB it is used which is not true.

Use __get_state() in these places instead to avoid this problem.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>