ranges: ignore certain parameter comparisons involving math
Say you have a function which does:

	return param + 100;

That gets stored in the DB as something like "100-199[==$0 + 100]".  The
problem is that Smatch was only looking at "==$0" and saying that it was
equal which is wrong.

The *right* fix would be to say that the return is ">$0" but my fix is to
just ignore it and put a note in the code that this is buggy.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>