ranges: fix generify_mtag_range()
The bug here is we shouldn't be using rl_intersection() at all.  Originally,
the idea was to say that if there is no intersection with the valid range
then we can return the original.  I guess I thought it would be a fast
path or something.  Anyway, delete that.

The other change is that I used the is_err_ptr() function instead of open
coding it in a pretty suspect looking way.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>