mtag_data: Use extra_mod hooks instead of modification hooks
The main problem with using smatch_modification_hooks.c was that when when
we have:

	struct foo *p = foo->bar->baz;

	p->a = 1;

Then we also need to set "foo->bar->baz->a = 1;".  This is handled in
Smatch extra but not by smatch_modification_hooks.c.  But the problem with
smatch extra was that it often was just using a var/sym pair instead of
an expression struct.  So I created gen_expression_from_name_sym() to
generate an expression.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>