type_val: ignore a bunch more kernel specific stuff
We want to ignore when structs are allocated to non-zero values because
hopefully that uninitialized data isn't used.  Before only kmalloc() was
handled but the kernel has a bunch of different ways to allocate data.

I also added a bunch of new ignored macros which hopefully don't change
unstructured buffers to structs.

The final change is that I made it ignore driver data because this is
very common.  This function does some caching so repeated look ups are
faster.  Intuitively, it felt like that would be a useful speed up.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>